# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Yingyan Platform API
  version: 0.1.0
  description: |
    映研用户端、管理端与 Worker 共用的 HTTP 契约。
    所有 JSON 成功响应均为 `{ code: 0, data }`，失败响应均为
    `{ code, message, details? }`。素材地址均为短期签名地址。
servers:
  - url: http://127.0.0.1:8080
    description: Local API
...
paths:
...
  /api/prompts/optimize:
    post:
      tags: [Generation]
      operationId: optimizePrompt
      security:
        - userSession: []
      requestBody:
        $ref: "#/components/requestBodies/PromptOptimize"
      responses:
        "200":
          $ref: "#/components/responses/PromptVersion"
        "400":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/prompts/reference-prompt:
    post:
      tags: [Generation]
      operationId: describeReferencePrompt
      security:
        - userSession: []
      requestBody:
        $ref: "#/components/requestBodies/ReferencePrompt"
      responses:
        "200":
          $ref: "#/components/responses/ReferencePrompt"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
  /api/prompts/confirm:
    post:
      tags: [Generation]
      operationId: confirmPrompt
      security:
        - userSession: []
      requestBody:
        $ref: "#/components/requestBodies/PromptConfirm"
      responses:
        "200":
          $ref: "#/components/responses/PromptVersion"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/generations:
    post:
      tags: [Generation]
      operationId: createGeneration
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/GenerationCreate"
      responses:
        "202":
          $ref: "#/components/responses/GenerationCreate"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/retouch-tickets/{ticketId}/quote/accept:
    post:
      tags: [Retouch]
      operationId: acceptRetouchQuote
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/TicketId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/RetouchQuoteAccept"
      responses:
        "200":
          $ref: "#/components/responses/RetouchWithEntitlement"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    userSession:
      type: apiKey
      in: cookie
      name: yy_user_session
    adminSession:
      type: apiKey
      in: cookie
      name: yy_manage_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    TicketId:
      name: ticketId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
...
    PromptOptimize:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PromptOptimizeInput"
    ReferencePrompt:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ReferencePromptInput"
    PromptConfirm:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PromptConfirmInput"
    GenerationCreate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenerationCreateInput"
...
    RetouchQuoteAccept:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [quoteId]
            properties:
              quoteId: { type: string, format: uuid }
...
  responses:
...
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    PromptVersion:
      description: 提示词版本
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PromptVersionSuccess"
...
    GenerationCreate:
      description: 创建生图任务并返回服务端最新权益
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenerationCreateSuccess"
...
    RetouchWithEntitlement:
      description: 工单与最新权益
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RetouchWithEntitlementSuccess"
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: integer, minimum: 1 }
        message: { type: string }
        details: true
...
    Entitlement:
      type: object
      additionalProperties: false
      required: [balance, canCreate, status]
      properties:
        balance: { type: integer, minimum: 0 }
        canCreate: { type: boolean }
        status: { type: string, enum: [unredeemed, active, empty] }
...
    ReferenceRole:
      type: string
      enum: [style, composition, person, detail]
    Asset:
      type: object
      additionalProperties: false
      required: [id, name, kind, mimeType, size, uploadProgress]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string }
        role:
          $ref: "#/components/schemas/ReferenceRole"
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        size: { type: integer, minimum: 0 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        previewUrl: { type: string, format: uri }
        previewUrlExpiresAt: { type: string, format: date-time }
        uploadProgress: { type: integer, minimum: 0, maximum: 100 }
        ownerId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...
    GenerationMode:
      type: string
      enum: [text-to-image, image-to-image]
    PromptSections:
      type: object
      additionalProperties: false
      required: [subject, scene, style, composition, details, negative, output]
      properties:
        subject: { type: string }
        scene: { type: string }
        style: { type: string }
        composition: { type: string }
        details: { type: string }
        negative: { type: string }
        output: { type: string }
        referencePrompt: { type: string, maxLength: 4000 }
    PromptReferenceAsset:
      type: object
      additionalProperties: false
      required: [assetId, role]
      properties:
        assetId: { type: string, format: uuid }
        role:
          $ref: "#/components/schemas/ReferenceRole"
    PromptOptimizeInput:
      type: object
      additionalProperties: false
      required: [source, mode, sourceAssetIds, referenceAssets]
      properties:
        source: { type: string, minLength: 4, maxLength: 2000 }
        mode:
          $ref: "#/components/schemas/GenerationMode"
        sourceAssetIds:
          type: array
          maxItems: 8
          uniqueItems: true
          items: { type: string, format: uuid }
        referenceAssets:
          type: array
          maxItems: 8
          items:
            $ref: "#/components/schemas/PromptReferenceAsset"
        referencePrompt: { type: string, maxLength: 4000 }
    ReferencePromptInput:
      type: object
      additionalProperties: false
      required: [referenceAssets]
      properties:
        referenceAssets:
          type: array
          minItems: 1
          maxItems: 8
          items:
            $ref: "#/components/schemas/PromptReferenceAsset"
...
    PromptConfirmInput:
      type: object
      additionalProperties: false
      required: [id, source, sections]
      properties:
        id: { type: string, format: uuid }
        source: { type: string, minLength: 1, maxLength: 2000 }
        sections:
          $ref: "#/components/schemas/PromptSections"
    PromptVersion:
      type: object
      additionalProperties: false
      required: [id, source, sections]
      properties:
        id: { type: string, format: uuid }
        source: { type: string }
        sections:
          $ref: "#/components/schemas/PromptSections"
        confirmedAt: { type: string, format: date-time }
    PromptVersionSuccess:
      type: object
      additionalProperties: false
      required: [code, data]
      properties:
        code: { type: integer, const: 0 }
        data:
          $ref: "#/components/schemas/PromptVersion"
...
    GenerationSettings:
      type: object
      additionalProperties: false
      required: [aspectRatio, outputCount, referenceStrength]
      properties:
        aspectRatio: { type: string, enum: ["1:1", "3:4", "4:3", "9:16", "16:9"] }
        outputCount: { type: integer, minimum: 1, maximum: 4 }
        referenceStrength: { type: integer, minimum: 0, maximum: 100 }
    GenerationCreateInput:
      type: object
      additionalProperties: false
      required: [promptVersionId, assetIds, settings]
      properties:
        promptVersionId: { type: string, format: uuid }
        assetIds:
          type: array
          maxItems: 8
          uniqueItems: true
          items: { type: string, format: uuid }
        settings:
          $ref: "#/components/schemas/GenerationSettings"
    GenerationStatus:
      type: string
      enum: [queued, processing, completed, partial, failed, cancelled]
    GenerationResult:
      type: object
      additionalProperties: false
      required: [id, url, downloadUrl, width, height]
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        downloadUrl:
          type: string
          format: uri
          description: 短期有效并带 attachment 文件名的签名下载地址
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
    GenerationTask:
      type: object
      required:
        - id
        - mode
        - title
        - status
        - prompt
        - settings
        - assets
        - requestedCount
        - successfulCount
        - reservedCredits
        - spentCredits
        - refundedCredits
        - progress
        - results
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        mode:
          $ref: "#/components/schemas/GenerationMode"
        title: { type: string }
        status:
          $ref: "#/components/schemas/GenerationStatus"
        prompt:
          $ref: "#/components/schemas/PromptVersion"
        settings:
          $ref: "#/components/schemas/GenerationSettings"
        assets:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        requestedCount: { type: integer, minimum: 1, maximum: 4 }
        successfulCount: { type: integer, minimum: 0, maximum: 4 }
        reservedCredits: { type: integer, minimum: 0 }
        spentCredits: { type: integer, minimum: 0 }
        refundedCredits: { type: integer, minimum: 0 }
        progress: { type: integer, minimum: 0, maximum: 100 }
        results:
          type: array
          items:
            $ref: "#/components/schemas/GenerationResult"
        retouchTicket:
          $ref: "#/components/schemas/RetouchSummary"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    GenerationCreateSuccess:
      type: object
      additionalProperties: false
      required: [code, data]
      properties:
        code: { type: integer, const: 0 }
        data:
          type: object
          additionalProperties: false
          required: [task, entitlement]
          properties:
            task:
              $ref: "#/components/schemas/GenerationTask"
            entitlement:
              $ref: "#/components/schemas/Entitlement"
...
    RetouchStatus:
      type: string
      enum:
        - submitted
        - quote_pending
        - accepted
        - processing
        - awaiting_confirmation
        - delivered
        - rejected
        - cancelled
    RetouchSummary:
      type: object
      required: [id, ticketNo, status, updatedAt]
      properties:
        id: { type: string, format: uuid }
        ticketNo: { type: string }
        status:
          $ref: "#/components/schemas/RetouchStatus"
        updatedAt: { type: string, format: date-time }
        quoteCredits: { type: integer, minimum: 1 }
...
    RetouchTicket:
      type: object
      required:
        - id
        - ticketNo
        - taskId
        - taskTitle
        - status
        - selectedResults
        - requirement
        - supplementalAssets
        - timeline
        - reservedCredits
        - spentCredits
        - refundedCredits
        - deliverables
        - sla
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        ticketNo: { type: string }
        taskId: { type: string, format: uuid }
        taskTitle: { type: string }
        status:
          $ref: "#/components/schemas/RetouchStatus"
        selectedResults:
          type: array
          items:
            $ref: "#/components/schemas/GenerationResult"
        requirement: { type: string }
        supplementalAssets:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        quote:
          type: object
          required: [id, credits, createdAt, status, expiresAt, remainingSeconds]
          properties:
            id: { type: string, format: uuid }
            credits: { type: integer, minimum: 1 }
            createdAt: { type: string, format: date-time }
            status: { type: string, enum: [active, accepted, invalidated, expired] }
            expiresAt: { type: string, format: date-time }
            remainingSeconds: { type: integer, minimum: 0 }
        timeline:
          type: array
          items:
            type: object
            required: [status, createdAt]
            properties:
              status:
                $ref: "#/components/schemas/RetouchStatus"
              action: { type: string }
              note: { type: string }
              createdAt: { type: string, format: date-time }
        reservedCredits: { type: integer, minimum: 0 }
        spentCredits: { type: integer, minimum: 0 }
        refundedCredits: { type: integer, minimum: 0 }
        revision:
          type: object
          properties:
            message: { type: string }
            requestedAt: { type: string, format: date-time }
        deliverables:
          type: array
          items:
            $ref: "#/components/schemas/GenerationResult"
        sla:
          $ref: "#/components/schemas/RetouchSLA"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RetouchSLA:
      type: object
      required: [stage, overdue]
      properties:
        stage: { type: string, enum: [quote, first-delivery, revision, completed] }
        dueAt: { type: [string, "null"], format: date-time }
        overdue: { type: boolean }
        remainingSeconds: { type: [integer, "null"], minimum: 0 }
...
    RetouchWithEntitlementSuccess:
      type: object
      additionalProperties: false
      required: [code, data]
      properties:
        code: { type: integer, const: 0 }
        data:
          type: object
          required: [ticket, entitlement]
          properties:
            ticket:
              $ref: "#/components/schemas/RetouchTicket"
            entitlement:
              $ref: "#/components/schemas/Entitlement"
...